Use matrix strategy at deploy github action
Use matrix strategy in `deploy.yml` github action to create and publish
package to `OpenVSX` and `Visual Studio Marketplace`.

It will use the following unique extension ids for registries:
- `OpenVSX`: `codechecker`
- `VS Marketplace`: `vscode-codechecker`

**Note**: unfortunately we can't use the same extension id for these two
registries because will get the following error message when the package
is published to the registry:
`The Extension Id already exist in the Marketplace. Please use the different Id`

Both package will be published under the `codechecker` organization.