Remove unclosable CodeChecker analysis in progress window
On large projects the analysis may take minutes. In this case it is
really annyoing for the user to show an unclosable pop-window.

Also this information whether an analysis is running is already available
in the bottom bar.

This patch will remove this pop-up window but will introduce a new tree
item on the overview sidebar to stop the analysis.